deleted add-curriculum component
- moved this into add-edit-curriculum screen due to multiple refactors
